check if order should be returned in amocrm

+++ src/amocrm/services/orders/order_returner.py
@@ -15,6 +15,17 @@ class AmoCRMOrderReturner(BaseService):
     order: Order
 
     def act(self) -> None:
+        if not self.order_must_be_returned():
+            return
+
         AmoCRMLead(order=self.order).update(status="closed")
         AmoCRMTransaction(order=self.order).delete()
         self.order.amocrm_transaction.delete()  # type: ignore
+
+    def order_must_be_returned(self) -> bool:
+        if self.order.is_b2b:
+            return False
+        if self.order.price == 0:
+            return False
+
+        return True
